Create Control.py
We create a RobotController class that subscribes to the /scan topic to receive laser scan data from the robot's sensors. We also create a publisher to send velocity commands to the robot on the /cmd_vel topic.

In the scan_callback method, we calculate the average distance to obstacles in the robot's field of view and adjust the robot's velocity based on this distance. If the average distance is greater than a desired distance (0.5 meters in this example), we set the linear velocity to 0.5 m/s and the angular velocity to 0.0 rad/s. If the average distance is less than the desired distance, we stop the robot by setting the linear velocity to 0.0 m/s and rotate it in place by setting the angular velocity to 0.5 rad/s.

You can also add more features like motion planning using algorithms such as A* or Dijkstra's algorithm. You can define a map of the environment and use the algorithm to find the shortest path to a target location while avoiding obstacles. Once the path is computed, you can convert it to a series of velocity commands to send to the robot's motors to make it move along the path.