Create remote.html
To run this code, make sure to replace the app.run line with the IP address of your Raspberry Pi or other robot platform. Then, run the code on the robot and open a web browser on your mobile device or computer and navigate to the IP address of the robot. You should see the remote control interface and be able to send velocity commands to the robot by filling out the input fields and clicking the "Move" button.